add js files
